updated broken front end tests and added a HouseholdMembersTable component test suite
