+++ app/javascript/components/patient/household/actions/ApplyToHousehold.js
@@ -8,7 +8,7 @@ class ApplyToHousehold extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      applyToHousehold: false
+      applyToHousehold: false,
     };
   }
