+++ app/javascript/tests/patient/Patient.test.js
@@ -6,6 +6,7 @@ import _ from 'lodash';
 import Patient from '../../components/patient/Patient';
 import BadgeHoH from '../../components/patient/household/utils/BadgeHoH';
 import { mockPatient1, mockPatient2, mockPatient3, mockPatient4, mockPatient5, blankMockPatient } from '../mocks/mockPatients';
+import { mockJurisdictionPaths } from '../mocks/mockJurisdiction';
 import { nameFormatter, formatDate } from '../util.js';
 
 const goToMock = jest.fn();
@@ -26,7 +27,7 @@ const riskFactors = [
 
 describe('Patient', () => {
   it('Properly renders all main components when not in edit mode', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#monitoree-details-header').exists()).toBeTruthy();
     expect(wrapper.find('#monitoree-details-header').find('h3').find('span').text()).toEqual(nameFormatter(mockPatient1));
     expect(wrapper.find('#monitoree-details-header').find(BadgeHoH).exists()).toBeTruthy();
@@ -45,7 +46,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders all main components when in edit mode', () => {
-    const wrapper = shallow(<Patient details={mockPatient4} goto={goToMock} collapse={true} edit_mode={true} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient4} goto={goToMock} collapse={true} edit_mode={true} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#monitoree-details-header').exists()).toBeTruthy();
     expect(wrapper.find('#monitoree-details-header').find('h3').find('span').text()).toEqual(nameFormatter(mockPatient4));
     expect(wrapper.find('#monitoree-details-header').find(BadgeHoH).exists()).toBeFalsy();
@@ -64,7 +65,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders identification section', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#identification');
     expect(section.find('h4').text()).toEqual('Identification');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -74,7 +75,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders contact information section', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#contact-information');
     expect(section.find('h4').text()).toEqual('Contact Information');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -84,7 +85,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders show/hide divider when props.collapse is true', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('.details-expander').exists()).toBeTruthy();
     expect(wrapper.find('#details-expander-link').exists()).toBeTruthy();
     expect(wrapper.find('.details-expander').find(FontAwesomeIcon).exists()).toBeTruthy();
@@ -94,7 +95,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders show/hide divider when props.collapse is false', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={false} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={false} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('.details-expander').exists()).toBeTruthy();
     expect(wrapper.find('#details-expander-link').exists()).toBeTruthy();
     expect(wrapper.find('.details-expander').find(FontAwesomeIcon).exists()).toBeTruthy();
@@ -104,7 +105,7 @@ describe('Patient', () => {
   });
 
   it('Clicking show/hide divider updates label and expands or collapses details', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find(Collapse).prop('in')).toBeFalsy();
     expect(wrapper.state('expanded')).toBeFalsy();
     wrapper.find('#details-expander-link').simulate('click');
@@ -116,7 +117,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders address section for domestic address with no monitoring address', () => {
-    const wrapper = shallow(<Patient details={mockPatient2} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient2} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#address');
     expect(section.find('h4').text()).toEqual('Address');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -130,7 +131,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders address section for domestic address and monitoring address', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#address');
     expect(section.find('h4').text()).toEqual('Address');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -150,7 +151,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders address section for foreign address with no monitoring address', () => {
-    const wrapper = shallow(<Patient details={mockPatient5} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient5} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#address');
     expect(section.find('h4').text()).toEqual('Address');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -164,7 +165,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders address section for foreign address and monitoring address', () => {
-    const wrapper = shallow(<Patient details={mockPatient4} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient4} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#address');
     expect(section.find('h4').text()).toEqual('Address');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -184,7 +185,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders arrival information section', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#arrival-information');
     expect(section.find('h4').text()).toEqual('Arrival Information');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -213,7 +214,7 @@ describe('Patient', () => {
   });
 
   it('Collapses/expands travel related notes if longer than 400 characters', () => {
-    const wrapper = shallow(<Patient details={mockPatient3} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient3} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#arrival-information').find('.notes-section').find(Button).exists()).toBeTruthy();
     expect(wrapper.state('expandArrivalNotes')).toBeFalsy();
     expect(wrapper.find('#arrival-information').find('.notes-section').find(Button).text()).toEqual('(View all)');
@@ -229,7 +230,7 @@ describe('Patient', () => {
   });
 
   it('Displays "None" if arrival information has no information', () => {
-    const wrapper = shallow(<Patient details={blankMockPatient} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={blankMockPatient} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#arrival-information');
     expect(section.exists()).toBeTruthy();
     expect(section.find('.none-text').exists()).toBeTruthy();
@@ -237,7 +238,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders planned travel section', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#planned-travel');
     expect(section.find('h4').text()).toEqual('Additional Planned Travel');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -252,7 +253,7 @@ describe('Patient', () => {
   });
 
   it('Collapses/expands additional planned travel notes if longer than 400 characters', () => {
-    const wrapper = shallow(<Patient details={mockPatient3} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient3} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#planned-travel').find('.notes-section').find(Button).exists()).toBeTruthy();
     expect(wrapper.state('expandPlannedTravelNotes')).toBeFalsy();
     expect(wrapper.find('#planned-travel').find('.notes-section').find(Button).text()).toEqual('(View all)');
@@ -268,7 +269,7 @@ describe('Patient', () => {
   });
 
   it('Displays "None" if planned travel has no information', () => {
-    const wrapper = shallow(<Patient details={blankMockPatient} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={blankMockPatient} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#planned-travel');
     expect(section.exists()).toBeTruthy();
     expect(section.find('.none-text').exists()).toBeTruthy();
@@ -276,7 +277,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders potential exposure information section', () => {
-    const wrapper = shallow(<Patient details={mockPatient2} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient2} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#potential-exposure-information');
     expect(section.find('h4').text()).toEqual('Potential Exposure Information');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -302,7 +303,7 @@ describe('Patient', () => {
   it('Properly renders notes in potential exposure information section if the rest of the section is empty', () => {
     let newMockPatient5 = _.cloneDeep(mockPatient5);
     newMockPatient5.exposure_notes = 'new exposure note';
-    const wrapper = shallow(<Patient details={newMockPatient5} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={newMockPatient5} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#potential-exposure-information');
     expect(section.find('.item-group').exists()).toBeFalsy();
     expect(section.find('.risk-factors').exists()).toBeFalsy();
@@ -313,7 +314,7 @@ describe('Patient', () => {
   });
 
   it('Collapses/expands exposure notes in potential exposure information section if longer than 400 characters', () => {
-    const wrapper = shallow(<Patient details={mockPatient5} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient5} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#potential-exposure-information').find(Button).exists()).toBeTruthy();
     expect(wrapper.find('#potential-exposure-information').find(Button).text()).toEqual('(View all)');
     expect(wrapper.find('#potential-exposure-information').find('.notes-text').text()).toEqual(mockPatient5.exposure_notes.slice(0, 400) + ' ...');
@@ -326,7 +327,7 @@ describe('Patient', () => {
   });
 
   it('Displays "None" if potential exposure information has no information', () => {
-    const wrapper = shallow(<Patient details={blankMockPatient} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={blankMockPatient} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#potential-exposure-information');
     expect(section.exists()).toBeTruthy();
     expect(section.find('.none-text').exists()).toBeTruthy();
@@ -337,7 +338,7 @@ describe('Patient', () => {
   });
 
   it('Properly renders case information section', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#case-information');
     expect(section.find('h4').text()).toEqual('Case Information');
     expect(section.find('.edit-link').exists()).toBeTruthy();
@@ -350,12 +351,12 @@ describe('Patient', () => {
   });
 
   it('Hides case information section when monitoree is in the exposure workflow', () => {
-    const wrapper = shallow(<Patient details={mockPatient2} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient2} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#case-information').exists()).toBeFalsy();
   });
 
   it('Properly renders notes section', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#exposure-notes');
     expect(section.find('h4').text()).toEqual('Notes');
     expect(section.find('.none-text').exists()).toBeFalsy();
@@ -365,7 +366,7 @@ describe('Patient', () => {
   });
 
   it('Collapses/expands exposure notes if longer than 400 characters', () => {
-    const wrapper = shallow(<Patient details={mockPatient3} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient3} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('#exposure-notes').find(Button).exists()).toBeTruthy();
     expect(wrapper.find('#exposure-notes').find(Button).text()).toEqual('(View all)');
     expect(wrapper.find('#exposure-notes').find('.notes-text').text()).toEqual(mockPatient3.exposure_notes.slice(0, 400) + ' ...');
@@ -378,7 +379,7 @@ describe('Patient', () => {
   });
 
   it('Displays "None" if exposure notes is null', () => {
-    const wrapper = shallow(<Patient details={mockPatient4} collapse={true} edit_mode={false} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient4} collapse={true} edit_mode={false} jurisdiction_paths={mockJurisdictionPaths} />);
     const section = wrapper.find('#exposure-notes');
     expect(section.exists()).toBeTruthy();
     expect(section.find('.none-text').exists()).toBeTruthy();
@@ -393,7 +394,7 @@ describe('Patient', () => {
   });
 
   it('Renders edit buttons if props.goto is defined', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} goto={goToMock} collapse={true} edit_mode={true} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} goto={goToMock} collapse={true} edit_mode={true} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('.edit-link').find(Button).length).toEqual(7);
     expect(wrapper.find('.edit-link').find('a').exists()).toBeFalsy();
     wrapper
@@ -406,7 +407,7 @@ describe('Patient', () => {
 
   it('Renders edit hrefs if props.goto is not defined', () => {
     const stepIds = [0, 2, 1, 3, 4, 5, 5];
-    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={true} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} collapse={true} edit_mode={true} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(wrapper.find('.edit-link').find(Button).exists()).toBeFalsy();
     expect(wrapper.find('.edit-link').find('a').length).toEqual(7);
     wrapper
@@ -419,7 +420,7 @@ describe('Patient', () => {
   });
 
   it('Calls props goto method when the edit buttons are clicked', () => {
-    const wrapper = shallow(<Patient details={mockPatient1} goto={goToMock} collapse={true} edit_mode={true} jurisdiction_path="USA, State 1, County 2" />);
+    const wrapper = shallow(<Patient details={mockPatient1} goto={goToMock} collapse={true} edit_mode={true} jurisdiction_paths={mockJurisdictionPaths} />);
     expect(goToMock).toHaveBeenCalledTimes(0);
     wrapper
       .find('.edit-link')
