+++ app/javascript/tests/patient/household/Dependent.test.js
@@ -2,30 +2,20 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { Row } from 'react-bootstrap';
 import Dependent from '../../../components/patient/household/Dependent';
+import HouseholdMemberTable from '../../../components/patient/household/utils/HouseholdMemberTable';
 import RemoveFromHousehold from '../../../components/patient/household/actions/RemoveFromHousehold';
 import { mockPatient1, mockPatient2 } from '../../mocks/mockPatients';
-import { nameFormatter } from '../../util.js';
+import { mockUser1 } from '../../mocks/mockUsers';
+import { mockJurisdictionPaths } from '../../mocks/mockJurisdiction';
 
 const mockToken = 'testMockTokenString12345';
 
 describe('Dependent', () => {
   it('Properly renders all main components', () => {
-    const wrapper = shallow(<Dependent patient={mockPatient2} hoh={mockPatient1} authenticity_token={mockToken} />);
-    expect(wrapper.find(Row).length).toEqual(2);
-    expect(wrapper.find(Row).at(0).text().includes('The reporting responsibility for this monitoree is handled by')).toBeTruthy();
-    expect(wrapper.find('a').exists()).toBeTruthy();
-    expect(wrapper.find('a').text()).toEqual(nameFormatter(mockPatient1));
-    expect(wrapper.find('a').prop('href').includes(`/patients/${mockPatient2.responder_id}`)).toBeTruthy();
-    expect(wrapper.find(RemoveFromHousehold).exists()).toBeTruthy();
-  });
-
-  it('Properly renders all main components if HoH is not defined', () => {
-    const wrapper = shallow(<Dependent patient={mockPatient2} authenticity_token={mockToken} />);
-    expect(wrapper.find(Row).length).toEqual(2);
-    expect(wrapper.find(Row).at(0).text().includes('The reporting responsibility for this monitoree is handled by')).toBeTruthy();
-    expect(wrapper.find('a').exists()).toBeTruthy();
-    expect(wrapper.find('a').text()).toEqual('this monitoree');
-    expect(wrapper.find('a').prop('href').includes(`/patients/${mockPatient2.responder_id}`)).toBeTruthy();
+    const wrapper = shallow(<Dependent patient={mockPatient2} other_household_members={[mockPatient1]} current_user={mockUser1} jurisdiction_paths={mockJurisdictionPaths} authenticity_token={mockToken} />);
+    expect(wrapper.find(Row).length).toEqual(3);
+    expect(wrapper.find(Row).at(0).text()).toEqual('This monitoree is a member of the following household:');
+    expect(wrapper.find(HouseholdMemberTable).exists()).toBeTruthy();
     expect(wrapper.find(RemoveFromHousehold).exists()).toBeTruthy();
   });
 });
